Rewrite calcium regulation (cooperativity) implementation
This new code should have mostly the same semantics as the old code. The
main difference is that the new code actually tests the cooperativity
conditions, so that e.g. cooperativity from XB3 but not TF3 could be
applied to some particular RU in one time step. As far as I can tell,
this is either just as correct (if only one can happen anyway) or
slightly more correct.